requirements.txt
flask>=2.0.0
apiflask>=0.4.0
flask-cors>=3.0.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.5
kubernetes>=26.1.0
pyyaml>=6.0
jinja2>=3.0.0
gunicorn>=20.1.0
python-dotenv>=0.19.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-flask>=1.2.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
marshmallow==3.21.3